% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_accessors.R
\docType{methods}
\name{advanced_parameters}
\alias{advanced_parameters}
\alias{advanced_parameters<-}
\title{Get and set Bobro class advanced parameters}
\usage{
advanced_parameters(x)

advanced_parameters(x) <- value
}
\arguments{
\item{x}{A \linkS4class{Bobro} class object.}

\item{values}{A string with same length of advancedparameters}
}
\value{
A list of basic parameters used in Bobro
}
\description{
Get and set Bobro class advanced parameters, default values are listed.
}
\details{

}
\examples{
advanced_parameters(object)
advanced_parameters(object) <-
}
\author{
Cankun Wang
}
