% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{BobroCommand-class}
\alias{BobroCommand-class}
\title{The BobroCommand Class}
\description{
The BobroCommand is used for logging commands that are run on a BobroObject. It stores parameters and timestamps
}
\section{Slots}{

\describe{
\item{\code{name}}{Command name}

\item{\code{time_stamp}}{Timestamp of when command was tun}

\item{\code{call_string}}{String of the command call}

\item{\code{params}}{List of parameters used in the command call}
}}

