% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_sequence.R
\docType{methods}
\name{compare_sequence}
\alias{compare_sequence}
\alias{compare_sequence,Bobro-method}
\title{Compare DNA sequence}
\usage{
\S4method{compare_sequence}{Bobro}(x)
}
\arguments{
\item{x}{A \linkS4class{Bobro} class object.}
}
\value{
A \linkS4class{Bobro} class object.
}
\description{
Compare DNA sequences using Bobro two-stage sequence alignment algorithm
}
\details{
Check \code{help_parameters} for current used parameters, to change parameters, use \code{basic_parameters()} and \code{advanced_parameters()}
}
\examples{
object <- compare_sequence(object)
}
\author{
Cankun Wang
}
