% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{CreateBobroObject}
\alias{CreateBobroObject}
\title{Create a Bobro object}
\usage{
CreateBobroObject(DNAStringObject, project = "BobroProject", meta_data = NULL)
}
\arguments{
\item{project}{Sets the project name for the Bobro object.}

\item{meta_data}{Additional metadata to add to the Bobro object.}
}
\description{
Create a Bobro object from a DNA string object (usually read fasta format file from package Biostrings).
}
\examples{
DNAStringObject <- readDNAStringSet("data/test1.fa")
my.object <- CreateBobroObject(DNAStringObject)
my.object
}
