% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{FindBobroMotif}
\alias{FindBobroMotif}
\title{Find Bobro motif}
\usage{
FindBobroMotif(object)
}
\arguments{
\item{project}{Sets the project name for the Bobro object.}
}
\description{
Find Bobro motif
}
\examples{
test_sequence <- Biostrings::readDNAStringSet("../data/cra.fa")
object <- CreateBobroObject(test_sequence)
object <- FindBobroMotif(object)

}
