% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_accessors.R
\name{help_parameters}
\alias{help_parameters}
\title{Display current parameters and description}
\usage{
help_parameters(x)
}
\arguments{
\item{x}{A \linkS4class{Bobro} class object.}
}
\description{
To change parameters, check the function \code{basic_parameters()} and \code{advanced_parameters()}
}
\details{

}
\examples{
help_parameters(object)
}
\author{
Cankun Wang
}
