% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{PlotBobroMotif}
\alias{PlotBobroMotif}
\title{Plot motif logo}
\usage{
PlotBobroMotif(object)
}
\arguments{
\item{project}{Sets the project name for the Bobro object.}
}
\description{
Plot motif logo
}
\examples{
test_sequence <- Biostrings::readDNAStringSet("../data/cra.fa")
object <- CreateBobroObject(test_sequence)
object <- FindBobroMotif(object)
PlotBobroMotif(object)

}
