% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_accessors.R
\docType{methods}
\name{sequence_names}
\alias{sequence_names}
\alias{sequence_names<-}
\title{Get and set DNA sequence headers}
\usage{
sequence_names(x)

sequence_names(x) <- value
}
\arguments{
\item{x}{A \linkS4class{Bobro} class object.}

\item{values}{A string with same length of names}
}
\value{
A character vector of DNA sequences sequence header information.
}
\description{
Get and set DNA sequence headers, default values read from FASTA format headers
}
\details{

}
\examples{
sequence_names(object)
sequence_names(object) <- paste("test-",sequence_names(object),sep="")
}
\seealso{
\code{\link{names}}, on which this function is based.
}
\author{
Cankun Wang
}
