% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{Bobro-class}
\alias{Bobro-class}
\title{The Bobro Class}
\description{
The Bobro object is a representation of DNA sequences and ChIP assay data for R; The
object was designed to be as self-contained as possible, and easily extendible to new methods.
}
\section{Slots}{

\describe{
\item{\code{assays}}{A list of assays for this project}

\item{\code{meta_data}}{Contains meta-information about each cell, starting with number of genes detected (nGene)
and the original identity class (orig.ident); more information is added using \code{AddMetaData}}

\item{\code{graphs}}{A list of \code{\link{Graph-class}} objects}

\item{\code{neighbors}}{...}

\item{\code{approximations}}{A list of matrix approximations for this object}

\item{\code{project.name}}{Name of the project}

\item{\code{misc}}{A list of miscellaneous information}

\item{\code{version}}{Version of Bobro this object was built under}

\item{\code{commands}}{A list of logged commands run on this \code{Bobro} object}

\item{\code{tools}}{A list of miscellaneous data generated by other tools, should be filled by developers only using \code{\link{Tool}<-}}

\item{\code{parameters}}{A list of parameters run on this \code{Bobro} object}
}}

